fixing location of testutil tests
